Netlink: MPLS routes in kernel, not fully working yet.
TODO: fix static to detect MPLS stack changes on reload

Anyway, Bird is now capable to insert both MPLS routes and MPLS encap
routes into kernel.

It was (among others) needed to define platform-specific AF_MPLS to 28
as this constant has been assigned in the linux kernel.

No support for BSD now, it may be added in the future.